+++ src/App.tsx
@@ -6,6 +6,7 @@ import { AppShell } from './components/layout/AppShell'
 import { TrackerProvider } from './context/TrackerProvider'
 import { GitHubSyncProvider } from './context/GitHubSyncProvider'
 import { ElaraProvider } from './context/ElaraProvider'
+import { TutorialProvider } from './context/TutorialProvider'
 import { ToastProvider } from './context/ToastProvider'
 import type { TabId } from './types'
 
@@ -20,7 +21,9 @@ function App() {
             <GitHubSyncProvider>
               <InstallReturnVisitSync />
               <ElaraProvider>
-                <AppShell activeTab={activeTab} onTabChange={setActiveTab} />
+                <TutorialProvider activeTab={activeTab} onTabChange={setActiveTab}>
+                  <AppShell activeTab={activeTab} onTabChange={setActiveTab} />
+                </TutorialProvider>
               </ElaraProvider>
             </GitHubSyncProvider>
           </TrackerProvider>